% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\alias{agg}
\alias{add_unit_id}
\alias{make_group}
\alias{compute_group}
\alias{compute_indicator_mean}
\alias{compute_indicator_median}
\alias{get_indicator_level}
\title{Aggregate data}
\usage{
agg(df, org, indicator)

add_unit_id(df, org)

make_group(df, group = "")

compute_group(gdf, description)

compute_indicator_mean(grouped_data)

compute_indicator_median(grouped_data)

get_indicator_level(grouped_data, description)
}
\arguments{
\item{df}{Data frame}

\item{org}{Data frame holding the org db table}

\item{indicator}{Data frame holding the indicator db table. Used as
description in some functions (please clean up)}

\item{group}{Character string defining grouping variable}

\item{gdf}{Data frame of grouped data}

\item{description}{see above}

\item{grouped_data}{see above}
}
\value{
Data frame in raw, grouped og aggregated form
}
\description{
Details here
}
