% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_no_dg}
\alias{check_no_dg}
\title{Replace "Ingen" with NA in a string}
\usage{
check_no_dg(s)
}
\arguments{
\item{s}{string}
}
\description{
This function is intended for use in indicator_server.
We would like the select_dg_id input to return NA,
but shiny::selectInput turns NAs into strings.
The input should therefore go through this function.
}
