% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{Aar}{Numeric 4 digit year}
  \item{ShNavn}{String hospital name}
  \item{ReshId}{String internal id of hospital}
  \item{OrgNrShus}{Numeric 9 digit id of hospital}
  \item{Variabel}{String variable name}
  \item{kvalIndID}{Numeric id of quality indicator}
  \item{delivery_id}{Numeric id referencing the delivery}
  }
}
\usage{
data
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
