% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\alias{make_pool}
\alias{insert_tab}
\alias{get_data}
\alias{drain_pool}
\title{Database functions for imongr}
\usage{
make_pool()

drain_pool(pool)

insert_tab(pool, table, df)

get_data(pool)
}
\arguments{
\item{pool}{a database connection pool object}

\item{table}{string defining target database table}

\item{df}{data frame containing data to be inserted into a database}
}
\value{
Database pool object, data frame or status message
}
\description{
Database credentials are read from config. If one or more of these are
defined 'env' corresponding values will be fetched from the environmental
variables IMONGR_DB_USER and IMONGR_DB_PASS. If these are undefined the
function providing connection handles will exit with an error.
}
