% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delivery}
\alias{delivery}
\title{Delivery example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{latest}{Numeric definig if the delivery is the latest (1) or not(0)}
  \item{latest_update}{Date (YYYY-MM-DD) defining data latest update}
  \item{latest_affirm}{Date (YYYY-MM-DD) up until data are regarded as final}
  \item{md5_checksum}{String A hash of the delivered data set used for
  integrity checking}
  \item{terms_version}{Character string providing version stamp of the terms
  accepted when data are published}
  \item{user_id}{Numeric id referencing the user that made the delivery}
  }
}
\usage{
delivery
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
