% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getenv.R
\name{getenv}
\alias{getenv}
\alias{db_host}
\alias{db_name}
\alias{db_username}
\alias{db_password}
\alias{adminer_url}
\title{Get relevant settings either from config og environmental variables}
\usage{
db_host()

db_name()

db_username()

db_password()

adminer_url()
}
\value{
Character string with settings
}
\description{
Get relevant settings either from config og environmental variables
}
