% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind}
\alias{ind}
\title{Indicator example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{id}{}
  \item{title}{}
  \item{name}{}
  \item{level_green}{}
  \item{level_yellow}{}
  \item{level_direction}{}
  \item{short_description}{}
  \item{long_description}{}
  \item{registry_id}{}

}
}
\usage{
ind
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
