% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator}
\alias{indicator}
\title{Indicator example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{IndId}{}
  \item{Register}{}
  \item{IndTittel}{}
  \item{IndNavn}{}
  \item{MaalNivaaGronn}{}
  \item{MaalNivaaGul}{}
  \item{MaalRetn}{}
  \item{BeskrivelseKort}{}
  \item{BeskrivelseLang}{}

}
}
\usage{
indicator
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
