% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicator.R
\name{levels_consistent_check}
\alias{levels_consistent_check}
\title{Check that the accomplishment thresholds are conistent}
\usage{
levels_consistent_check(input, conf)
}
\arguments{
\item{input}{Shiny input object}

\item{conf}{get_config() output}
}
\description{
This function is intended for use in indicator_server.
Threshold are set for high accomplishment (green)
and medium accomplihment (yellow). The green threshold
should be higher than the yellow. If not, and error message
is displayed.
}
