% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{insert_sample_data}
\alias{delete_all_data}
\title{Tools and whatever}
\usage{
insert_sample_data(include_data_table = TRUE)

delete_all_data()
}
\arguments{
\item{include_data_table}{Logical defining if the data table is to be
populated by data too. By default TRUE}
}
\value{
Invisible
}
\description{
Tools and whatever
}
