% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicator.R, R/mod_project.R
\name{mod_indicator}
\alias{mod_indicator}
\alias{indicator_ui}
\alias{indicator_server}
\alias{indicator_app}
\title{Shiny module providing GUI and server logic for the indicator tab}
\usage{
indicator_ui(id)

indicator_server(id, registry_tracker, pool, pool_verify)

indicator_app(pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{registry_tracker}{Integer defining registry id}

\item{pool}{A database pool object}

\item{pool_verify}{A database pool object}
}
\value{
Shiny objects for the imongr app

Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for the indicator tab

Shiny module providing GUI and server logic for the indicator tab
}
