% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_notice.R
\name{mod_notice}
\alias{mod_notice}
\alias{notice_ui}
\alias{notice_server}
\alias{notice_app}
\title{Shiny module providing GUI and server logic for the notice tab}
\usage{
notice_ui(id)

notice_server(id, registry_tracker, pool, pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{registry_tracker}{Integer defining registry id}

\item{pool}{A database pool object}

\item{pool_verify}{A database pool object}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for the notice tab
}
