% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profile.R
\name{mod_profile}
\alias{mod_profile}
\alias{profile_ui}
\alias{profile_server}
\alias{profile_app}
\title{Shiny module providing GUI and server logic for (user) profile}
\usage{
profile_ui(id)

profile_server(id, pool, pool_verify)

profile_app(pool, pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{pool}{A database pool object connecting to production data}

\item{pool_verify}{A database pool object connecting to staging data}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for (user) profile
}
