% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_publication.R
\name{mod_publication}
\alias{mod_publication}
\alias{publication_ui}
\alias{publication_server}
\alias{publication_app}
\title{Shiny module providing GUI and server logic for the publication tab}
\usage{
publication_ui(id)

publication_server(id, registry_tracker, pool, pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{registry_tracker}{Integer defining registry id}

\item{pool}{A database pool object}

\item{pool_verify}{A database pool object}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for the publication tab
}
