% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report}
\alias{mod_report}
\alias{report_ui}
\alias{report_server}
\alias{report_app}
\title{Shiny module providing GUI and server logic for reports}
\usage{
report_ui(id)

report_server(id, pool, pool_verify)

report_app(pool, pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{pool}{A database pool object connecting to production data}

\item{pool_verify}{A database pool object connecting to staging data}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for reports
}
