% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_selected_indicators.R
\name{mod_selected_indicators}
\alias{mod_selected_indicators}
\alias{selected_indicators_ui}
\alias{selected_indicators_server}
\title{Shiny module providing GUI and server logic for the selected indicators tab}
\usage{
selected_indicators_ui(id)

selected_indicators_server(id, registry_tracker, pool, pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{registry_tracker}{Integer defining registry id}

\item{pool}{A database pool object}

\item{pool_verify}{A database pool object}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for the selected indicators tab
}
