% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_status.R
\name{mod_status}
\alias{mod_status}
\alias{status_ui}
\alias{status_server}
\alias{status_app}
\alias{add_arrows}
\title{Shiny module providing UI and server functions for registry status overview}
\usage{
add_arrows(dat_format)

status_ui(id)

status_server(id, pool, pool_verify)

status_app(pool, pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{pool}{A database pool object connecting to production data}

\item{pool_verify}{A database pool object connecting to staging data}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing UI and server functions for registry status overview
}
