% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload}
\alias{mod_upload}
\alias{upload_ui}
\alias{upload_server}
\alias{upload_app}
\title{Shiny module providing GUI and server logic for the upload data tab}
\usage{
upload_ui(id)

upload_server(id, registry_tracker, pool_verify)

upload_app(pool)
}
\arguments{
\item{id}{Character string module namespace}

\item{registry_tracker}{Integer defining registry id}

\item{pool_verify}{A database pool object}

\item{pool}{A database pool object}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for the upload data tab
}
