% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{ops}
\alias{ops}
\alias{delete_indicator_data}
\alias{delete_registry_data}
\alias{delete_agg_data}
\alias{insert_data}
\alias{insert_agg_data}
\alias{update_aggdata_delivery}
\alias{agg_all_data}
\alias{clean_agg_data}
\alias{create_imongr_user}
\alias{update_registry_medfield}
\alias{update_registry_user}
\alias{update_ind_text}
\alias{update_ind_val}
\alias{insert_data_verify}
\alias{insert_data_prod}
\title{Functions for data operations in imongr}
\usage{
delete_indicator_data(pool, df)

delete_agg_data(pool, df)

insert_data_verify(pool, df, update = NA, affirm = NA)

insert_data_prod(
  pool_verify,
  pool_prod,
  df,
  registry_delivery_ids,
  terms_version = NA
)

insert_agg_data(pool, df)

update_aggdata_delivery(pool, indicator)

agg_all_data(pool)

clean_agg_data(pool)

create_imongr_user(pool, df)

update_registry_medfield(pool, registry_id, df)

update_registry_user(pool, registry_id, df)
}
\arguments{
\item{pool}{Database pool object}

\item{df}{A data frame with new user data for overwriting the old user list}

\item{update}{Character string of format YYYY-MM-DD providing date until data
are regarded as updated. Default value is NA.}

\item{affirm}{Character string of format YYYY-MM-DD providing date until data
are regarded as final. Default value is NA.}

\item{pool_verify}{A database pool object}

\item{pool_prod}{A database pool object}

\item{registry_delivery_ids}{Integer delivery ids}

\item{terms_version}{Character string providing version stamp of of the terms
accepted when data are published. Default value is NA that will normally
apply for all uploads prior to publishing.}

\item{indicator}{Character vector of indicator ids}

\item{registry_id}{The numeric id of the selected registry}
}
\value{
Relevant values from the current environment and database
}
\description{
Functions for data operations in imongr
}
