% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/org.R
\docType{data}
\name{org}
\alias{org}
\alias{create_imongr_org}
\alias{create_imongr_user}
\title{Organization example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{name}{String with name of organization. The person (user) adding
  data to qmonger must belong to an organization found in \code{org$name}}
  \item{orgnumber}{Numeric id of organization}
  \item{valid}{Numeric defining if organization is active (1) of not (0)}
}
}
\usage{
org

create_imongr_org(pool, name, orgnumber, valid = 1)

create_imongr_user(pool, user_name, name, phone, email, valid = 1, orgnumber)
}
\arguments{
\item{pool}{Database connection pool object}

\item{name}{String defining any given name to the entity}

\item{orgnumber}{Integer providing a global id of the entity}

\item{valid}{Integer defining if the entity is valid (1) or deprecated (0).
Default value is 1}

\item{user_name}{String defining the username db field. Most likely its
content should be unique}

\item{phone}{String providing the phone number of a user}

\item{email}{String providing the e-mail of a user}
}
\value{
These functions are mainly fronends to DBI and the returned value is
mainly what will be porvided by the underlying DBI functions
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
