% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orgnr}
\alias{orgnr}
\title{Organization structure example data}
\format{
Data frame with example data
\describe{
  \item{orgnr}{}
  \item{unit_level}{}
}
}
\usage{
orgnr
}
\description{
Flat org hierarcy. To be removed...
}
\keyword{datasets}
