% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orgnr_shortname}
\alias{orgnr_shortname}
\title{Organization structure short_name example data}
\format{
Data frame with example data
\describe{
  \item{orgnr}{}
  \item{unit_level}{}
  \item{short_name}{}
}
}
\usage{
orgnr_shortname
}
\description{
All orgs with unit_level and short_names.
}
\keyword{datasets}
