% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicator.R
\name{oversize_check}
\alias{oversize_check}
\title{Display an oversize warning if there are too many characters in the input text}
\usage{
oversize_check(oversize, conf)
}
\arguments{
\item{oversize}{Logical}

\item{conf}{get_config() output}
}
\description{
This function is intended for use in indicator_server.
If the input text for indicator title, short description
or long description is too long according to the specification
in the config file, then an error message is displayed.
}
