% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publish}
\alias{publish}
\title{publish example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{md5_checksum}{}
  \item{terms_version}{}
  \item{user_id}{}
  \item{registry_id}{}
}
}
\usage{
publish
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refer to the overall data model (ref)
}
\keyword{datasets}
