% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registry}
\alias{registry}
\title{Registry example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{Register}{String with registry name}
  \item{full_name}{String with registry full name}
}
}
\usage{
registry
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
