% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registry_medfield}
\alias{registry_medfield}
\title{registry_medfield example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{id}{}
  \item{registry_id}{}
  \item{medfield_id}{}
}
}
\usage{
registry_medfield
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refer to the overall data model (ref)
}
\keyword{datasets}
