% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\alias{registry_status_report}
\alias{.registry_status_data}
\title{Functions providing reports to be used in imongr}
\usage{
.registry_status_data(pool)

registry_status_report(pool, pool_verify)
}
\arguments{
\item{pool}{Database pool object for connection to the production system}

\item{pool_verify}{Database pool object for connection to staged data}
}
\value{
A data frame holding the report data
}
\description{
Functions providing reports to be used in imongr
}
