% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhf}
\alias{rhf}
\title{Organization structure example data}
\format{
Data frame with example data
\describe{
  \item{id}{}
  \item{orgnr}{}
  \item{full_name}{}
  \item{short_name}{}
  \item{ref_id}{}
}
}
\usage{
rhf
}
\description{
Org hierarcy
}
\keyword{datasets}
