% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{user}
\alias{user}
\title{User example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{user_name}{String user name of the person altering data, \emph{e.g.}
  'jesusc'}
  \item{user}{String full name of the person altering data, \emph{e.g.}
  'Jesus Christ'}
  \item{phone}{String phone number of the person altering data, \emph{e.g.}
  '+4747474747}
  \item{email}{String e-mail address of the person altering data,
  \emph{e.g.} 'jesusc\@sky.com'}
  \item{valid}{Numeric definig if user is active (1) or not (0)}
  }
}
\usage{
user
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
