% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{user_registry}
\alias{user_registry}
\title{User_registry example data for qmongr data store}
\format{
Data frame with made up values
\describe{
  \item{Register}{}
  \item{user_id}{}
}
}
\usage{
user_registry
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
