% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\alias{agg}
\alias{add_unit_id}
\alias{make_group}
\alias{compute_group}
\alias{compute_indicator_mean}
\alias{compute_indicator_median}
\alias{get_indicator_level}
\title{Aggregate data}
\usage{
agg(df, org, indicator)

add_unit_id(df, org)

make_group(df, group = "")

compute_group(gdf, indicator)

compute_indicator_mean(gdf)

compute_indicator_median(gdf)

get_indicator_level(gdf, indicator)
}
\arguments{
\item{df}{Data frame}

\item{org}{Data frame holding the org db table providing all organization
data for each unit (from hospital to health trust)}

\item{indicator}{Data frame holding the indicator db table providing all
data on each indicator}

\item{group}{Character string defining the name of the grouping variable
in a data frame}

\item{gdf}{Data frame of grouped data}
}
\value{
Data frame in raw, grouped or aggregated form
}
\description{
Aggregation of indicators as provided for each unit (hospital) at
organization levels above hospital: health trust, regional health trust and
national.
}
\details{
All functions are adapted from qmongr/qmongrdata. However, the nomenclature
of function arguments are changed somewhat. Main source of underlying data
is the qmongr database and where tables are used in these function their
names are kept as is (\emph{org} and \emph{indicator}). Other data frames
passed into or between function are denoted \emph{df} or \emph{gdf} where
the latter denotes a data frame that has been grouped (by
\code{dplyr::group_by()}.
}
