% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\alias{agg}
\alias{agg_from_level}
\alias{agg_residual}
\alias{agg_udef}
\alias{make_group}
\alias{compute_group}
\alias{compute_indicator_mean}
\alias{compute_indicator_median}
\alias{get_indicator_level}
\title{Aggregate hospital data}
\usage{
agg(df, org, ind)

agg_from_level(df, org, ind, conf, from_level)

agg_residual(aggs, conf)

make_group(df, group = "")

compute_group(gdf, ind)

compute_indicator_mean(gdf)

compute_indicator_median(gdf)

get_indicator_level(gdf, ind)
}
\arguments{
\item{df}{Data frame}

\item{org}{Data frame holding the flat org table with all levels needed. May
be obtained by \code{get_flat_org(pool)}}

\item{ind}{Data frame holding the indicator db table providing all
data on each indicator}

\item{group}{Character string defining the name of the grouping variable
in a data frame}

\item{gdf}{Data frame of grouped data}
}
\value{
Data frame in raw, grouped or aggregated form
}
\description{
Aggregation of indicators as provided for each unit (hospital) at
organization levels above hospital: health trust, regional health trust and
national.
}
\details{
All functions are adapted from qmongr/qmongrdata. However, the nomenclature
of function arguments are changed somewhat. Main source of underlying data
is the qmongr database and where tables are used in these function their
names are kept as is (\emph{org} and \emph{ind}). Other data frames
passed into or between function are denoted \emph{df} or \emph{gdf} where
the latter denotes a data frame that has been grouped (by
\code{dplyr::group_by()}. The aggregate consists of the following varables:
\describe{
  \item{year}{The year of the current record}
  \item{ind_id}{Indicator ID}
  \item{orgnr}{Organization ID}
  \item{count}{Number of observations of the current record}
  \item{var}{Summarised indicator value, for instance mean or median}
  \item{level}{Code providing evaluated discret level such as 'H' (high),
  'M' (intermediate) and 'L' (low)}
  \item{level_direction}{Name providing the desired level such as 'Lavt'
  (low) or 'Høyt' (high)}
  \item{unit_level}{Code provideing level of aggregation such as 'hospital',
  'hf' (health trust), 'rhf' (regional health trust) and 'national'}
  \item{unit_name}{Name of the organization unit}
}
}
