% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\title{Functions for checks in imongr}
\arguments{
\item{input}{A shiny input object}

\item{conf}{The data from the get_config function}

\item{ns}{A shiny::NS namespace function}

\item{rv}{A shiny::reactiveValues object}
}
\description{
Functions for checks in imongr
}
