% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{create_config}
\alias{get_config}
\alias{check_config}
\alias{write_config}
\title{Functions handling imongr R package config}
\usage{
create_config(dir = ".")

write_config(config, dir = ".", filename = "_imongr.yml")

get_config(dir = ".")

check_config(config)
}
\arguments{
\item{dir}{string providing path to config}

\item{config}{list containing config}
}
\value{
A status message or list of config
}
\description{
Functions handling imongr R package config
}
