% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\alias{make_pool}
\alias{drain_pool}
\alias{insert_tab}
\alias{get_data}
\alias{get_delivery}
\alias{get_user}
\alias{get_user_registry}
\alias{get_table}
\alias{get_org}
\title{Low level database functions for imongr}
\usage{
make_pool()

drain_pool(pool)

insert_tab(pool, table, df)

get_table(pool, tab)

get_data(pool)

get_delivery(pool)

get_user(pool)

get_org(pool)

get_user_registry(pool)
}
\arguments{
\item{pool}{a database connection pool object}

\item{table}{string defining target database table}

\item{df}{data frame containing data to be inserted into a database}
}
\value{
Database pool object, data frame or status message
}
\description{
Database metadata are read from config. If one or more of these are
defined 'env' corresponding values will be fetched from the environmental
variables IMONGR_DB_HOST, IMONGR_DB_NAME, IMONGR_DB_USER and IMONGR_DB_PASS.
If these are undefined the function providing connection handles will exit
with an error. What the function expects from table names and variable
names in the data is also defined by the configuration. Thus, functions can
be left unchanged when the data model is altered as long as such changes
are reflected by configuration.
}
\details{
Ordianry user interactions with data should have their own functions, but
may be built ontop of these ones. For instance, such functions must make
sure consistency (\emph{e.g.} foreign keys) between database tables are
kept in order.
}
