% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get.R
\name{db_get}
\alias{db_get}
\alias{get_registry_name}
\alias{get_org_name}
\title{Retreiv data from imongr database}
\usage{
get_registry_name(pool, registry, full_name = FALSE)

get_org_name(pool, orgnr)
}
\arguments{
\item{pool}{a database connection pool object}

\item{registry}{Integer defining registry id}

\item{full_name}{Logical defining if full names is to be returned}

\item{orgnr}{Integer id of organization}
}
\value{
Data object from database
}
\description{
Database metadata are read from config. If one or more of these are
defined 'env' corresponding values will be fetched from the environmental
variables IMONGR_DB_HOST, IMONGR_DB_NAME, IMONGR_DB_USER and IMONGR_DB_PASS.
If these are undefined the function providing connection handles will exit
with an error. What the function expects from table names and variable
names in the data is also defined by the configuration. Thus, functions can
be left unchanged when the data model is altered as long as such changes
are reflected by configuration.
}
