% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgmt.R
\name{mgmt}
\alias{mgmt}
\alias{create_imongr_org}
\alias{create_imongr_user}
\title{Functions for organization and user management in imongr}
\usage{
create_imongr_org(pool, df)

create_imongr_user(pool, df)
}
\arguments{
\item{pool}{Database connection pool object}

\item{df}{Data frame corresponding to database table fields}
}
\description{
Functions for organization and user management in imongr
}
