% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{navbar_widget}
\alias{version_info}
\alias{no_opt_out_ok}
\alias{insert_sample_data}
\alias{delete_all_data}
\title{Tools and whatever}
\usage{
navbar_widget()

version_info(newline = "<br>")

no_opt_out_ok()

insert_sample_data(include_data_table = TRUE)

delete_all_data(prompt = TRUE)
}
\arguments{
\item{newline}{String element defining line break for formatting. Default is
\code{<br>}}

\item{include_data_table}{Logical defining if the data table is to be
populated by data too. By default TRUE}

\item{prompt}{Logical to prompt for user input. Default is TRUE}
}
\value{
Invisible
}
\description{
Tools and whatever
}
\examples{
version_info()
no_opt_out_ok()
}
