% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{natural}
\alias{md5_checksum}
\alias{user_widget}
\alias{version_info}
\alias{no_opt_out_ok}
\alias{insert_sample_data}
\alias{delete_all_data}
\alias{remove_empty_rows}
\alias{invalidate_cache}
\alias{validateName}
\title{Tools and whatever}
\usage{
natural(vals, tolerance = .Machine$double.eps^0.5)

md5_checksum(df, ind = "")

remove_empty_rows(df)

user_widget()

version_info(newline = "<br>")

no_opt_out_ok()

insert_sample_data(include_data_table = TRUE)

delete_all_data(prompt = TRUE)

invalidate_cache()

validateName(x, existing_names)
}
\arguments{
\item{vals}{A numeric vector}

\item{tolerance}{A (small) positive floating point number used to evaluate if
a numeric can be regarded as a whole number. Default depends on the running
environment and set to \code{.Machine$double.eps^0.5}}

\item{df}{A data frame}

\item{ind}{A data frame holding indicator data.}

\item{newline}{String element defining line break for formatting. Default is
\code{<br>}}

\item{include_data_table}{Logical defining if the data table is to be
populated by data too. By default TRUE}

\item{prompt}{Logical to prompt for user input. Default is TRUE}
}
\value{
Invisible
}
\description{
Tools and whatever

Function for removing empty rows in csv input.
In a row ";;;;;" the values will translate to
NA or an empty string depending on whether the columns
are numeric or strings.
}
\examples{
version_info()
no_opt_out_ok()
}
