% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicator.R
\name{mod_indicator}
\alias{mod_indicator}
\alias{indicator_input}
\alias{indicator_server}
\alias{indicator_app}
\title{Shiny module providing GUI and server logic for the indicator tab}
\usage{
indicator_input(id)

indicator_server(id, pool)

indicator_app(pool)
}
\arguments{
\item{id}{Character string module namespace}

\item{pool}{A database pool object}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for the indicator tab
}
