% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_publish.R
\name{mod_publish}
\alias{mod_publish}
\alias{publish_ui}
\alias{publish_server}
\alias{publish_app}
\title{Shiny module providing GUI and server logic for the publish indicator tab}
\usage{
publish_ui(id)

publish_server(id, tab_tracker, registry_tracker, pool, pool_verify)

publish_app(pool, pool_verify)
}
\arguments{
\item{id}{Character string module namespace}

\item{tab_tracker}{String defining tab}

\item{registry_tracker}{Integer defining registry id}

\item{pool}{A database pool object}

\item{pool_verify}{A database pool object}
}
\value{
Shiny objects for the imongr app
}
\description{
Shiny module providing GUI and server logic for the publish indicator tab
}
