% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{ops}
\alias{ops}
\alias{get_user_data}
\alias{get_user_id}
\alias{get_user_latest_delivery_id}
\alias{get_registry_data}
\alias{md5_checksum}
\alias{delivery_exist_in_db}
\alias{retire_user_deliveries}
\alias{delete_registry_data}
\alias{insert_data}
\alias{get_all_user_data}
\alias{get_user_registries}
\alias{get_user_deliveries}
\title{Functions for data operations in imongr}
\usage{
get_all_user_data(pool)

get_user_data(pool)

get_user_id(pool)

get_user_registries(pool)

get_user_deliveries(pool)

get_user_latest_delivery_id(pool)

get_registry_data(pool, registry)

md5_checksum(df)

delivery_exist_in_db(pool, df)

retire_user_deliveries(pool)

delete_registry_data(pool, df)

insert_data(pool, df)
}
\arguments{
\item{pool}{Database connection pool object}

\item{registry}{Character registry name}

\item{df}{Data frame of relevant data}
}
\value{
Relevant values from the current environment and database
}
\description{
Functions for data operations in imongr
}
