% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{ops}
\alias{ops}
\alias{delivery_exist_in_db}
\alias{duplicate_delivery}
\alias{retire_user_deliveries}
\alias{delete_indicator_data}
\alias{delete_registry_data}
\alias{delete_agg_data}
\alias{insert_data}
\alias{insert_agg_data}
\alias{agg_all_data}
\alias{clean_agg_data}
\alias{create_imongr_user}
\title{Functions for data operations in imongr}
\usage{
delivery_exist_in_db(pool, df)

duplicate_delivery(pool, df)

retire_user_deliveries(pool)

delete_indicator_data(pool, df)

delete_agg_data(pool, df)

insert_data(pool, df)

insert_agg_data(pool, df)

agg_all_data(pool)

clean_agg_data(pool)

create_imongr_user(pool, df)
}
\arguments{
\item{pool}{Database connection pool object}

\item{df}{Data frame of relevant data}
}
\value{
Relevant values from the current environment and database
}
\description{
Functions for data operations in imongr
}
