% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{ops}
\alias{ops}
\alias{delivery_exist_in_db}
\alias{duplicate_delivery}
\alias{retire_user_deliveries}
\alias{delete_indicator_data}
\alias{delete_registry_data}
\alias{delete_agg_data}
\alias{insert_data}
\alias{insert_agg_data}
\alias{update_aggdata_delivery}
\alias{update_aggdata_delivery_time}
\alias{agg_all_data}
\alias{clean_agg_data}
\alias{create_imongr_user}
\alias{update_registry_medfield}
\alias{update_registry_user}
\alias{update_ind_text}
\title{Functions for data operations in imongr}
\usage{
delivery_exist_in_db(pool, df)

duplicate_delivery(pool, df)

retire_user_deliveries(pool)

delete_indicator_data(pool, df)

delete_agg_data(pool, df)

insert_data(pool, df, update = NA, affirm = NA, terms_version = NA)

insert_agg_data(pool, df)

update_aggdata_delivery(pool)

update_aggdata_delivery_time(pool)

agg_all_data(pool)

clean_agg_data(pool)

create_imongr_user(pool, df)

update_registry_medfield(pool, df)

update_registry_user(pool, df)

update_ind_text(pool, df)
}
\arguments{
\item{pool}{Database connection pool object}

\item{df}{Data frame of relevant data}

\item{update}{Character string of format YYYY-MM-DD providing date until data
are regarded as updated. Default value is NA.}

\item{affirm}{Character string of format YYYY-MM-DD providing date until data
are regarded as final. Default value is NA.}

\item{terms_version}{Character string providing version stamp of of the terms
accepted when data are published. Default value is NA that will normally
apply for all uploads prior to publishing.}
}
\value{
Relevant values from the current environment and database
}
\description{
Functions for data operations in imongr
}
