% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{org}
\alias{org}
\title{Organization example data for qmongr data store}
\format{Data frame with made up values
\describe{
  \item{name}{String with name of organization. The person (user) adding
  data to qmonger must belong to an organization found in \code{org$name}}
  \item{orgnumber}{Numeric id of organization}
  \item{valid}{Numeric defining if organization is active (1) of not (0)}
}}
\usage{
org
}
\description{
Meta data to allow interactions with qmong data. Only used for testing in
this package. Please refere to the overall data model (ref)
}
\keyword{datasets}
