% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{org}
\alias{org}
\title{Organization structure example data}
\format{
Data frame with example data
\describe{
  \item{hospital_id}{}
  \item{hospital}{}
  \item{orgnr_hospital}{}
  \item{hf}{}
  \item{orgnr_hf}{}
  \item{rhf}{}
  \item{orgnr_rhf}{}
  \item{national}{}
  \item{orgnr_national}{}
}
}
\usage{
org
}
\description{
Flat org hierarcy
}
\keyword{datasets}
