% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_output.R
\name{server_output}
\alias{server_output}
\alias{select_registry_ui}
\alias{submit_ui}
\alias{error_report_ui}
\alias{upload_sample_text_ui}
\alias{upload_sample_ui}
\alias{var_doc_ui}
\alias{profile_ui}
\title{Functions that provide server output}
\usage{
profile_ui(conf, pool, valid_user, iusr, igrs)

select_registry_ui(conf, pool)

submit_ui(conf, pool, upload_file, registry, df)

error_report_ui(pool, df, upload_file, registry)

upload_sample_text_ui(conf, upload_file)

upload_sample_ui(df, upload_file, registry, sample_size, sample_type)

var_doc_ui(conf)
}
\arguments{
\item{conf}{List of configuration}

\item{pool}{Data base pool object}

\item{valid_user}{Logical valid user}

\item{iusr}{Character string username}

\item{igrs}{Character string comma separated user groups}

\item{upload_file}{Character string with file to upload}

\item{registry}{Character string definig registry}

\item{df}{Data frame}

\item{sample_size}{Numeric sample size}

\item{sample_type}{Character string of sampling type}
}
\value{
shiny ui objects to be provided by the shiny server function
}
\description{
Functions that provide server output
}
