% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_output.R
\name{server_output}
\alias{server_output}
\alias{select_registry_ui}
\alias{submit_ui}
\alias{error_report_ui}
\alias{upload_sample_text_ui}
\alias{upload_sample_ui}
\alias{var_doc_ui}
\alias{medfield_summary_text_ui}
\alias{reguser_summary_text_ui}
\alias{profile_ui}
\title{Functions that provide server output}
\usage{
profile_ui(conf, pool, valid_user, iusr, igrs)

select_registry_ui(pool, conf, input_id, context, current_reg = character())

submit_ui(conf, pool, upload_file, registry, df, context)

error_report_ui(pool, df, upload_file, registry)

upload_sample_text_ui(pool, conf, upload_file, registry, indicators)

upload_sample_ui(df, upload_file, registry, sample_size, sample_type)

var_doc_ui(conf)
}
\arguments{
\item{conf}{List of configuration}

\item{pool}{Data base pool object}

\item{valid_user}{Logical valid user}

\item{iusr}{Character string username}

\item{igrs}{Character string comma separated user groups}

\item{input_id}{Character string with shiny ui id}

\item{context}{Character string with user selected (db) context}

\item{current_reg}{Character string defining previously selected registry}

\item{upload_file}{Character string with file to upload}

\item{registry}{Character string defining registry}

\item{df}{Data frame}

\item{indicators}{Character vector of indicators}

\item{sample_size}{Numeric sample size}

\item{sample_type}{Character string of sampling type}
}
\value{
shiny ui objects to be provided by the shiny server function
}
\description{
Functions that provide server output
}
