% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload}
\alias{upload}
\alias{check_report}
\alias{check_upload}
\alias{check_missing_registry}
\alias{check_missing_var}
\alias{check_invalid_var}
\alias{check_invalid_org}
\alias{check_invalid_context}
\alias{check_invalid_ind}
\alias{check_none_numeric_var}
\alias{check_duplicate_delivery}
\alias{csv_to_df}
\alias{mail_check_report}
\alias{sample_df}
\alias{indicator_is_fraction}
\alias{check_numeric_var}
\alias{check_natural_var}
\alias{check_overflow_var}
\alias{check_numeric_denominator}
\alias{check_natural_denominator}
\alias{check_zero_denominator}
\title{Upload data to imongr}
\usage{
check_report(registry, df, pool)

mail_check_report(pool, registry, mail_msg)

check_upload(registry, df, pool)

check_missing_registry(registry, df, conf, pool)

check_missing_var(registry, df, conf, pool)

check_invalid_var(registry, df, conf, pool)

check_invalid_context(registry, df, conf, pool)

check_invalid_org(registry, df, conf, pool)

check_invalid_ind(registry, df, conf, pool)

check_numeric_var(registry, df, conf, pool)

check_natural_var(registry, df, conf, pool)

check_overflow_var(registry, df, conf, pool)

check_numeric_denominator(registry, df, conf, pool)

check_natural_denominator(registry, df, conf, pool)

check_zero_denominator(registry, df, conf, pool)

check_duplicate_delivery(registry, df, conf, pool)

csv_to_df(path, sep = ",", dec, encoding = "UTF-8")

sample_df(df, skip = c(""), n, random = FALSE)

indicator_is_fraction(pool, df, conf, return_ind = FALSE)
}
\arguments{
\item{registry}{Integer registry id}

\item{df}{Data frame}

\item{pool}{Data base pool object}

\item{mail_msg}{Character vector holding (part of) email body}

\item{conf}{List of configuration}

\item{path}{Character path to a file}

\item{sep}{Character filed sep}

\item{dec}{Character decimal sep}

\item{encoding}{Character encoding}

\item{skip}{character vector defining data frame variables to skip}

\item{n}{Numeric sample size}

\item{random}{Logical sample method}

\item{return_ind}{Logical whether indicators should be returned. FALSE by
default}
}
\value{
whatever
}
\description{
Various functions used when data are being uploaded to imongr. All checks
regarding consistency goes here
}
